+++ tools/release-plugins/info/lib/getInfo.ts
@@ -1,6 +1,7 @@
 import { mergeOptions } from '@tikkhun/utils-core';
 import { Info } from './info.interface';
 import { loadInfo } from './loadInfo/loadInfo.utils';
+import { getInfoFromNestedObject } from './utils';
 export type LoadInfoArgs = any[];
 export interface FromObject {
   prefix?: string;
@@ -20,59 +21,66 @@ export interface GetInfoOptions {
   // 可以添加前缀
   from: From[];
   debug?: boolean;
+  ignoreError?: boolean; // 是否忽略错误
 }
 export async function getInfo(options: GetInfoOptions): Promise<Info> {
   const opts = mergeOptions(
     {
       from: [],
       debug: false,
+      ignoreError: true,
     },
     options,
   ) as GetInfoOptions;
-  if (!opts.from.length) {
+  if (!opts.from?.length) {
     if (options.debug) {
       console.debug(`from is empty`);
     }
     return {};
   }
-  const infoSources: Info[] = await Promise.all(
-    opts.from.map(async (fromOptions) => {
-      // 要么 数组 要么 对象
-      if (typeof fromOptions !== 'object') {
-        return null;
+
+  const infoSources = await Promise.all(
+    opts.from.filter(isObject).map(normalizeFromOptions).map(async (fromOptions) => {
+      try {
+        const { prefix, args, map } = fromOptions || {};
+        // 没有参数就直接返回null
+        if (!args?.length) {
+          return null;
+        }
+        const originInfo = await loadInfo(...args);
+        return transInfo(originInfo, { prefix, map });
+      } catch (error) {
+        if (opts.ignoreError) {
+          console.error('获取信息失败,但被忽略:', error);
+          return {};
+        }
+        throw error;
       }
-      if (Array.isArray(fromOptions)) {
-        return loadInfo(...fromOptions);
-      }
-      const { prefix, args, map } = fromOptions || {};
-      // 没有参数就直接返回null
-      if (!args?.length) {
-        return null;
-      }
-      const originInfo = await loadInfo(...args);
-      const mappedInfo = isEmpty(map) ? originInfo : getMappedInfo(originInfo, map!);
-      // 这里可以设置prefix
-      const values = prefix ? { [prefix]: mappedInfo } : mappedInfo;
-      return values;
     }),
   );
   return mergeOptions(...infoSources); // 合并
 }
-
-export function getMappedInfo(data: any, map: Record<string, string>) {
-  // 数据不是data
-  if (typeof data !== 'object') {
-    return data;
+export function isObject(v: any) {
+  return typeof v === 'object';
+}
+/**
+ * @function normalizeFromOptions
+ * @description 归一化from选项,如果是数组,则转换为对象
+ * @param options
+ * @returns
+ */
+export function normalizeFromOptions(options: From | FromObject): FromObject {
+  if (Array.isArray(options)) {
+    return {
+      args: options,
+    }
   }
-  const mappedInfo: Record<string, any> = {};
-  for (const [key, value] of Object.entries(map)) {
-    mappedInfo[key] = data[value];
+  return options;
+}
+export function transInfo(origin: any, { prefix, map }: { prefix?: string, map?: Record<string, string> }) {
+  const mappedInfo = getInfoFromNestedObject(origin, map);
+  if (prefix) {
+    return { [prefix]: mappedInfo };
   }
   return mappedInfo;
-}
-export function isEmpty(obj: any) {
-  if (!obj) {
-    return true;
-  }
-  return Object.keys(obj).length === 0;
 }