+++ tools/release-plugins/info/lib/index.ts
@@ -1,4 +1,4 @@
 export * from './loadInfo';
 export * from './getInfo';
 export * from './info.interface';
-
+export * from './utils'