+++ tools/release-plugins/info/lib/loadInfo/index.ts
@@ -1,2 +1,1 @@
 export * from "./loadInfo.utils";
-export * from "./getInfoFromNestedObject";