+++ tools/release-plugins/info/src/index.ts
@@ -1,12 +1,19 @@
-import { getInfo } from '../lib';
+import { fileURLToPath } from 'url';
+import { __FILE_STAT__, getInfo } from '../lib';
+
 async function bootstrap() {
   const info = await getInfo({
     from: [
+      // [
+      //   {
+      //     name: 'alibaba',
+      //   },
+      // ],
+      // ['package.json'],
       [
-        {
-          name: 'alibaba',
-        },
-      ] as any,
+        __FILE_STAT__,
+        process.cwd()
+      ]
     ],
   });
   console.log(`info`, info);